services/report-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.2
celery==5.4.0
redis==5.0.4
httpx==0.27.0
reportlab==4.2.0
pandas==2.2.2
python-multipart==0.0.9
supabase==2.4.0
pika==1.3.2
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-asyncio==0.23.0
